test(rule): move helpers out of tests/lib
